my-zulip-cert-hook: Custom certificate hook
Custom certificate handling hook.  Assumes that landfill is managing the
certificates using certbot or similar.  When the script runs it checks
the actual certificates on the webserver to see if they are due to
expire in less than 29 days.  If so, it copies the certs from landfill
and then into the docker container and updates the appropriate services.

Signed-off-by: Bill Pittman <bill.pittman@ni.com>